feat(pages): simplify feed page structure
