--- client/src/pages/feed/index.xml.ts
@@ -1,3 +1,0 @@
-import { Feed } from './_shared';
-
-export const GET = await Feed();
